サーバーサイドでメール処理したいのでFlaskで作った。DBもサーバ側に移管すると思う。ユーザに対してメールアドレス登録などをⅰ一意にしたいのでログイン機能つけることにした。
