+++ .gitignore
@@ -6,3 +6,5 @@ __pycache__
 token.pickle
 sent_message_id.txt
 forms_id.txt
+venv
+database.db
