+++ backend/models.py
@@ -1,0 +1,20 @@
+import sqlite3
+
+def create_connection():
+    conn = sqlite3.connect('database.db')
+    return conn
+
+def create_table():
+    conn = create_connection()
+    with conn:
+        conn.execute('''
+            CREATE TABLE IF NOT EXISTS users (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            email TEXT NOT NULL UNIQUE,
+            password TEXT NOT NULL
+        )
+        ''')
+    conn.close()
+
+# アプリケーションの開始時にテーブルを作成
+create_table()
