+++ resource_manager/resource_manager.py
